% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_m1.R
\name{adaBoost}
\alias{adaBoost}
\title{adaBoost}
\usage{
adaBoost(formula, data, nIter, ...)
}
\arguments{
\item{formula}{Formula for models}

\item{data}{Input dataframe}

\item{nIter}{no. of classifiers}
}
\description{
Implements Freund's Adaboost.M1 algorithm
}
\examples{
fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
test.adaboost <- adaBoost(Y~X, data=fakedata,10)
}

