% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_fast.R
\name{adaboost_fast}
\alias{adaboost_fast}
\title{fast implementation of adaboost using Rcpp}
\usage{
adaboost_fast(formula, data, nIter, ...)
}
\arguments{
\item{formula}{Formula for models}

\item{data}{Input dataframe}

\item{nIter}{no. of classifiers}
}
\description{
fast implementation of adaboost using Rcpp
}

