% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{returns an individual tree from the adaboost object}
\usage{
get_tree(object, tree_num)
}
\arguments{
\item{object}{object of class fastBoost}

\item{tree_num}{integer describing the tree to get}
}
\value{
tree object of class rpart
}
\description{
returns an individual tree from the adaboost object
}
\examples{
fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
A <- adaboost(Y~X, fakedata, 10)
tree <- get_tree(A,5)
}

