% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_fast.R
\name{precondition_checks}
\alias{precondition_checks}
\title{checks if inputs are in an expected format
and stops right away if something is wrong}
\usage{
precondition_checks(formula, data)
}
\arguments{
\item{formula}{formula for models}

\item{data}{input dataframe}
}
\description{
checks if inputs are in an expected format
and stops right away if something is wrong
}

