% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.adaboost_m1.R
\name{predict.adaBoost}
\alias{predict.adaBoost}
\title{makes predictions for an adaboost object on a new dataset}
\usage{
\method{predict}{adaBoost}(object, newdata)
}
\arguments{
\item{object}{an object of class adaboost}

\item{newdata}{dataframe on which we are looking to predict}
}
\value{
predictor predcted object with prediction probabilities
}
\description{
makes predictions for an adaboost object on a new dataset
}
\examples{
A <- adaBoost(Y~X, fakedata, 10)
pred <- predict(A,newdata=fakedata)
}

