% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_adaboost.R
\name{predict.real_adaboost}
\alias{predict.real_adaboost}
\title{makes predictions for an real_adaboost object on a new dataset}
\usage{
\method{predict}{real_adaboost}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class real_adaboost}

\item{newdata}{dataframe on which we are looking to predict}

\item{...}{arguments passed to predict.default}
}
\value{
predictor predicted object with prediction probabilities
}
\description{
makes predictions for an real_adaboost object on a new dataset
}
\examples{
fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
A <- real_adaboost(Y~X, fakedata, 10)
pred <- predict(A,newdata=fakedata)
}

