% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_adaboost.R
\name{predict_adaboost_internal}
\alias{predict_adaboost_internal}
\title{internal method which predicts both adaboost_M1 and 
real adaboost}
\usage{
predict_adaboost_internal(object, newdata, method)
}
\arguments{
\item{object}{an object of class adaboost}

\item{newdata}{dataframe on which we are looking to predict}

\item{method}{M1 or SAMME.R}
}
\description{
internal method which predicts both adaboost_M1 and 
real adaboost
}

