% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adaboost.R
\name{print.adaboost}
\alias{print.adaboost}
\title{S3 method to print an adaboost object}
\usage{
\method{print}{adaboost}(x, ...)
}
\arguments{
\item{x}{object of class adaboost}

\item{...}{arguments passed to print.default}
}
\value{
None
}
\description{
S3 method to print an adaboost object
}
\examples{
fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
A <- adaboost(Y~X, fakedata, 10)
print(A)

}

