% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_real.R
\name{real_adaboost}
\alias{real_adaboost}
\title{Implements Zhu et al's real adaboost or SAMME.R algorithm}
\usage{
real_adaboost(formula, data, nIter, ...)
}
\arguments{
\item{formula}{Formula for models}

\item{data}{Input dataframe}

\item{nIter}{no. of classifiers}

\item{...}{other optional arguments, not implemented now}
}
\description{
Implements Zhu et al's real adaboost or SAMME.R algorithm
}
\examples{
fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
test.adaboost <- real_adaboost(Y~X, data=fakedata,10)
}

