% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_rpart.R
\name{wrap_rpart}
\alias{wrap_rpart}
\title{wraps the rpart call for ease of use with Rcpp}
\usage{
wrap_rpart(formula_obj, newdata, weight_vec, classnames_map)
}
\arguments{
\item{formula_obj}{formula object}

\item{newdata}{dataframe with requisite columns}

\item{weight_vec}{vector of numeric values,weights for each example}

\item{classnames_map}{named vector mapping classnames to 0/1.}
}
\description{
wraps the rpart call for ease of use with Rcpp
}

