% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_rpart_predict.R
\name{wrap_rpart_predict_real}
\alias{wrap_rpart_predict_real}
\title{wraps the prediction of probability from a fitted tree
for use with Rcpp}
\usage{
wrap_rpart_predict_real(tree_object, newdata)
}
\arguments{
\item{tree_object}{object of class rpart}

\item{newdata}{dataframe}
}
\value{
pred_vec vector of probability that
                example belongs to class 0
}
\description{
wraps the prediction of probability from a fitted tree
for use with Rcpp
}

