% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_fast.R
\name{adaboost_fast}
\alias{adaboost_fast}
\title{fast implementation of adaboost using Rcpp}
\usage{
adaboost_fast(formula, data, nIter, method)
}
\arguments{
\item{formula}{Formula for models}

\item{data}{Input dataframe}

\item{nIter}{no. of classifiers}

\item{method}{a string, either "M1" or "SAMME.R"}
}
\value{
adaboost_object An object of class adaboost/real_adaboost containing each tree, and its corresponding weight
}
\description{
fast implementation of adaboost using Rcpp
}

