\name{fastBoost-package}
\alias{fastBoost-package}
\alias{fastBoost}
\docType{package}
\title{
\packageTitle{fastBoost}
}
\description{
\packageDescription{fastBoost}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastBoost}
\packageIndices{fastBoost}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastBoost}

Maintainer: \packageMaintainer{fastBoost}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
