% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adaboost.R
\name{print.adaBoost}
\alias{print.adaBoost}
\title{S3 method to print an adaboost object}
\usage{
\method{print}{adaBoost}(object)
}
\arguments{
\item{object}{object of class adaboost}
}
\value{
None
}
\description{
S3 method to print an adaboost object
}
\examples{
#'fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
A <- adaBoost(Y~X, fakedata, 10)
print(A)

}

