% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_rpart.R
\name{wrap_rpart}
\alias{wrap_rpart}
\title{wraps the rpart call for ease of use with Rcpp}
\usage{
wrap_rpart(newdata)
}
\arguments{
\item{newdata}{dataframe with requisite columns}
}
\description{
wraps the rpart call for ease of use with Rcpp
}

