% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_rpart_predict.R
\name{wrap_rpart_predict}
\alias{wrap_rpart_predict}
\title{wraps the prediction from a tree for ease of use 
with Rcpp}
\usage{
wrap_rpart_predict(tree_object, newdata, classnames_map)
}
\arguments{
\item{tree_object}{object of class rpart}

\item{newdata}{dataframe}
}
\value{
pred_vec integer vector. This contains
       the labels of the predicted class 
       converted to integers.
}
\description{
wraps the prediction from a tree for ease of use 
with Rcpp
}

